[fix][broker] Allow Access to System Topic Metadata for Reader Creation Post-Namespace Deletion
## Motivation
After initiating the snapshot segment function, deletion of topics necessitates the activation of readers. Furthermore, these readers should be opened and deleted as they are used, which implies that we should not pre-store readers. However, after initiating the deletion of namespaces currently, it is not allowed to obtain the metadata of partition topics, making it impossible to create readers. This results in the inability to delete namespaces.

## Modification
Allow the acquisition of system topic metadata after initiating namespace deletion, thus creating readers to clean up topic data.