[fix][client] Make the whole grabCnx() progress atomic
### Motivation

In `ConnectionHandler`, there is a `Connection` interface whose methods
will be called after the connection in `grabCnx` is established, the
implementation of `Connection` might send some requests after the
connection is established. For example, the consumer will send the
`CommandSubscribe` request in `connectionOpened`. However, the whole
process is not atomic, which leads to the message lost reported in
https://github.com/apache/pulsar/pull/20591.

### Modifications

Modify the `Connection` interface to have a single method:

```java
CompletableFuture<Void> handleNewConnection(ClientCnx cnx, PulsarClientException e);
```

The returned future should be completed once the implementation has done
everything, e.g. for the consumer, the future should only be completed
after receiving the response for `CommandSubscribe`.

In `grabCnx`, the `ConnectionHandler` could only connect to the broker
once the whole process is completed.

Add `ConnectionHandlerTest` to verify the behavior.