[fix][broker] Avoid throwing RestException in BrokerService (#20761)
## Motivation
The BrokerService should not throw RestException, as the method is not only called by the admin tool. When a topic is auto-created during the creation of a consumer/producer and exceeds the maximum number of topics allowed in the namespace, returning a RestException (UnknownError received by the client) would be considered a retryable error. 
## Modification
To address this issue, the BrokerService will now return a NotAllowException instead of a RestException when the topic count exceeds the maximum limit allowed in the namespace.