Fix breaking change of the deprecated constructor of PersistentMessageExpiryMonitor
### Motivation

https://github.com/apache/pulsar/pull/20781 adds a new constructor to
`PersistentMessageExpiryMonitor` and initialize the old constructor with

```java
this.topic = subscription.topic;
```

NPE will happen when `subscription` is null. However, it's allowed to
pass a null `subscription` for test because methods like
`findEntryFailed` don't depend on the `topic` field.

### Modifications

Add the null check and mark the old constructor as deprecated.