Fix the built-in admin failed to delete a topic with a custom authz provider
### Motivation

https://github.com/apache/pulsar/pull/20496 introduced a breaking change
that if the custom authz provider does not implement the
`removePermissionsAsync` method, the built-in admin will fail to delete
a topic. It's because `BrokerService#deleteTopicAuthenticationWithRetry`
calls `authorizationService.removePermissionsAsync`.

### Modifications

To make it backward compatible, return a normal completed future in
`removePermissionsAsync`. Verify the change by deleting a topic in
`AuthorizationWithAuthDataTest#testAdmin`.