[fix] [ml] Reader can set read-pos to a deleted ledger (#21248)
### Motivation

After trimming ledgers, the variable `lastConfirmedEntry` of the managed ledger might rely on a deleted ledger(the latest ledger which contains data).

There is a bug that makes pulsar allow users to set the start read position to an unexisting ledger or a deleted ledger when creating a subscription. This makes the `backlog` and `markDeletedPosition` wrong.

### Modifications

Fix the bug.

(cherry picked from commit 4ee5cd78147890ce3c23092edbfc11370972728e)