[fix] [build] rename schema_example.conf to schema_example.json (#21447)
### Motivation

Since the files will be checked for `LICENSE` whose name is a suffix with `.conf`. But we can not add a LICENSE definition for the file `schema_example.conf` because it is a `JSON` file.

Note: the file `schema_example.conf` is only used for the integration test `CLITest.java`.

### Modifications

Rename `schema_example.conf` to `schema_example.json`.

(cherry picked from commit 517961e749120e8f8cb5217a56cf37bbb50183ec)