[improve][sec] Align the default mechanism for server to request certificates
The current default behavior of Netty and Jetty's SSLContext implementation is that the
server will request the client certificate, and the client can optionally pass it on.

[Netty SSL Context](https://github.com/apache/pulsar/blob/e1d06b5f54f08c09debab7a9a513b7c173c1779b/pulsar-common/src/main/java/org/apache/pulsar/common/util/SecurityUtility.java#L575)
[Jetty SSL Context](https://github.com/apache/pulsar/blob/e1d06b5f54f08c09debab7a9a513b7c173c1779b/pulsar-broker-common/src/main/java/org/apache/pulsar/jetty/tls/JettySslContextFactory.java#L106)

But the implementation of key store is inconsistent, it will lead server do not request
the client certificate and can not do authentication without `tlsRequireTrustedClientCertOnConnect`
configuration.