[improve][broker] PIP-300: Add custom dynamic configuration for plugins
### Motivation

See [PIP-300](https://github.com/apache/pulsar/pull/21127)

### Modifications

Implement `registerCustomDynamicConfiguration`, which just adds a
customized key to `dynamicConfigurationMap` so that updating such a key
will be possible. The value is a `ConfigField` object whose `Field`
field is null.

Add `testUpdateCustomDynamicConfiguration` to verify it works.

It should be noted the implementation is slightly different from the
original proposal, which also does not express the design clearly.