[fix][client] Fix messages in the batch container timed out unexpectedly
Fixes https://github.com/apache/pulsar/issues/21884

### Motivation

When `ProducerImpl#run` is called where `pendingMessages` is empty and
`batchMessageContainer` is not empty, the whole batch's timestamp is
treated as `lastBatchSendNanoTime + batch latency`.

Given send timeout as 5 seconds, assuming a batch was flushed at `t1`
and the next message was sent after 5 seconds:
- t1: `ProducerImpl#batchFlushTask()`, `lastBatchSendNanoTime` is now `t1`
- t1 + 0.0011 s: `ProducerImpl#run()`, no pending message and the batch container is empty
- t1 + 5.0010 s: `sendAsync()`, the message is added to batch
- t1 + 5.0011 s: `ProducerImpl#run()` is called before `batchFlushTask()`

https://github.com/apache/pulsar/blob/176bdeacd309e8c1e49358987a1946abd30ba34a/pulsar-client/src/main/java/org/apache/pulsar/client/impl/ProducerImpl.java#L2014-L2015

`createdAt` will be `t1 + 0.001 s`. However, the current time point is
`t1 + 5.0011 s`, the interval is 5.001 s, timeout happened.

### Modification

Record the timestamp when the 1st message is added to the batch
container and use this timestamp instead of `lastBatchSendNanoTime` to
compute the `createdAt`.

Add `testSendTimerCheckForBatchContainer` to cover this case.