[improve][txn] Add command to abort transaction (#21630)
### Motivation

<!-- Explain here the context, and why you're making that change. What is the problem you're trying to solve. -->
In the implementation of Pulsar  transaction, if we have a stuck transaction, then the transactions after this one cannot be consumed by the consumer even if they have been committed. The consumer will be stuck until the stuck transaction is aborted due to timeout, and then it will continue to consume messages. Therefore, we need to provide an admin command to proactively abort the transaction in such situations.
### Modifications

<!-- Describe the modifications you've done. -->
Introduce a new API for aborting transaction.