[fix][broker] Ignore non-retryable errors when loading topic-level policies before loading topics
### Motivation

https://github.com/apache/pulsar/pull/21445 introduces an implicit
requirement that if the topic policies cannot be loaded,
`BrokerService#getTopic` will fail and then the client will retry
loading the topic.

It could break some existing usages like the tests in the C++
client: https://github.com/apache/pulsar-client-cpp/pull/394

This change is applied only to avoid the race condition when unloading a
namespace bundle. However, the client should not fail if the topic-level
policies are not available.

### Modifications

If the topic policies cannot be loaded due to a non-retryable error, we
should not fail the `getTopic`. For retryable errors, the client will
still retry until the broker gets the topic-level policies successfully
after some attempts.

Modify `TokenAuthenticatedProducerConsumerTest` to protect the change.