[cleanup] [broker] fix return value of TransactionTimeoutTrackerImpl#addTransaction (#22020)
### Motivation
According to the method signature, we should return true when the transaction is added to the tracker.
```
    /**
     * Add a txnID to the tracker.
     *
     * @param sequenceId
     *            the sequenceId
     * @param timeout
     *            the absolute timestamp for transaction timeout
     *
     * @return true if the transaction was added to the tracker or false if had timed out
     */
    CompletableFuture<Boolean> addTransaction(long sequenceId, long timeout);
```
But actually, we return false for any cases.

Update: Moreover, we do not use the return value anyway, it is useless. We have better remove it.

### Modifications

~~Return true when the transaction is added to the tracker.~~
Remove the return value as it is useless.