[improve][admin] Expose the offload threshold in seconds to the amdin ---
### Motivation

We have two configurations to trigger the offload automatically,
time and bytes. We only expose the bytes in the admin but didn't
expose the time in the admin. When we want to disable it, we need
to set both of them to the -1. But we didn't expose the time settings,
the offload will never be disabled.