[fix][broker] Fix wrong double-checked locking for readOnActiveConsumerTask in dispatcher
### Motivation

The access on `readOnActiveConsumerTask` is not thread safe.

```java
if (readOnActiveConsumerTask != null) {
    return;
}
readOnActiveConsumerTask = topic.getBrokerService().executor().schedule(() -> {/* ... */});
```

There is a case that:

| Steps | Thread 1 | Thread 2 |
| :- | :- | :- |
| 1 | Read `readOnActiveConsumerTask`: null | |
| 2 | call `schedule` | |
| 3 | | Read `readOnActiveConsumerTask`: null |
| 4 | | call `schedule` |
| 5 | Write `readOnActiveConsumerTask` to the result of `schedule` | |
| 6 | | Write `readOnActiveConsumerTask` to the result of `schedule` |

Then `schedule()` will be called twice and only one result will be
assigned to `readOnActiveConsumerTask`.

### Modifications

Follow the double-checked locking when `readOnActiveConsumerTask` is
null to ensure `readOnActiveConsumerTask` cannot be called concurrently.