[improve][broker] Remove the atomicity on active consumer of a dispatcher
### Motivation

The modifications on
`AbstractDispatcherSingleActiveConsumer#activeConsumer` only happen on
`pickAndScheduleActiveConsumer`, which is only called in `addConsumer`
and `removeConsumer`. However, both `addConsumer` and `removeConsumer`
are synchronized methods so that we don't have to perform an atomic
operation on `activeConsumer`.

The composite operations on `activeConsumer` also happen on synchronized
methods like `disconnectActiveConsumers`. The only access to
`activeConsumer` that is not synchronized is the read on
`activeConsumer` like `getActiveConsumer()` and `isConsumerConnected()`.

### Modifications

Remove the `ACTIVE_CONSUMER_UPDATER` and note some methods are not
thread safe or there is no need to make them thread safe.

Besides, since `pickAndScheduleActiveConsumer` is never called
concurrently, we don't need a double-checked locking on
`readOnActiveConsumerTask`. Instead, just synchronize the modification
in the scheduled task.