[fix][client] Fix wrong results of hasMessageAvailable and readNext after seeking by timestamp
### Motivation

After seeking by timestamp, `hasMessageAvailable()` and `readNext()`
could return wrong values.

The 1st bug was a regression introduced by
https://github.com/apache/pulsar/pull/22201, which modifies
`startMessageId` to `seekMessageId` before a seek operation is done.
However, the previous behavior is also a bug but accidentally works in
this case. When seeking by timestamp, the `seekMessageId` is modified to
`earliest`, which should not be compared with `lastMessageIdInBroker`
because the actual start position is determined by the seek timestamp,
not the `earliest` message id.

The 2nd bug was caused by https://github.com/apache/pulsar/pull/9652,
when `startMessageIdInclusive()` is configured to create a reader, it
could seek to the position of the latest message when
`lastDequeuedMessageId` is `earliest` and `startMessageId` is `latest`.

### Modifications

Add a boolean flag `hasSoughtByTimestamp` to represent whether the last
seek call accepts a timestamp. If it's true, don't take `startMessageId`
into comparison with `lastMessageIdInBroker`, just compare the
mark-delete position and last position in the GetLastMessageId response.

Add `testHasMessageAvailableAfterSeekTimestamp` to verify the change.

For the `readNext` call, if the reader has sought by timestamp, don't
seek to the latest position in `hasMessageAvailable`. Modify
`testReadMessageWithBatchingWithMessageInclusive` to verify the fix.
However, this patch does not modify the existing behavior when `seek` is
not called because the inclusive reader relies on the implicit seek
operation in `hasMessageAvailable`.