[improve][broker] Make BrokerSelectionStrategy pluggable
### Motivation

When users want to extend the `ExtensibleLoadManagerImpl`, the
`BrokerSelectionStrategy` cannot be customized, so users have to write
much duplicated code for a customized broker selection strategy.

### Modifications

Add a stable interface `BrokerSelectionStrategyFactory` and implements
this interface in `ExtensibleLoadManagerImpl` with the default
implementation that returns a `LeastResourceUsageWithWeight` instance by
default. Add `CustomBrokerSelectionStrategyTest` to show how to
customize the broker selection strategy and verify it works.